+++ adaptavist/const.py
@@ -1,0 +1,5 @@
+KEEP_ORIGINAL_VALUE = r"\{keep_original_value\}"
+
+TEST_CASE = "TEST_CASE"
+TEST_PLAN = "TEST_PLAY"
+TEST_RUN = "TEST_RUN"