+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 
 - *BREAKING*: If the docs say list, they mean list. Strings are not supported anymore.
+- *BREAKING*: Do move test cases to the root folder, an empty string is needed
 
 ## [1.1.2]